chore: forcing HMR to work;
