+++ src/components/ui/toggle.tsx
@@ -36,4 +36,5 @@ const Toggle = React.forwardRef<
 
 Toggle.displayName = TogglePrimitive.Root.displayName;
 
+// eslint-disable-next-line react-refresh/only-export-components
 export { Toggle, toggleVariants };
